+++ src/cpx_io/cpx_system/cpx_ap/cpx_ap_module_definitions.py
@@ -10,6 +10,8 @@ from cpx_io.cpx_system.cpx_ap.ap8do import CpxAp8Do
 from cpx_io.cpx_system.cpx_ap.ap12di4do import CpxAp12Di4Do
 from cpx_io.cpx_system.cpx_ap.ap16di import CpxAp16Di
 from cpx_io.cpx_system.cpx_ap.vabx_ap import VabxAP
+from cpx_io.cpx_system.cpx_ap.vaem_ap import VaemAP
+from cpx_io.cpx_system.cpx_ap.vmpal_ap import VmpalAP
 
 # Dict that maps from module ids to corresponding module classes
 CPX_AP_MODULE_ID_DICT = {
@@ -29,6 +31,8 @@ CPX_AP_MODULE_ID_DICT = {
     "LKC": CpxAp4Iol,
     "LKB": CpxAp4Iol,
     "VABX": VabxAP,
+    "VAEM": VaemAP,
+    "VMPAL": VmpalAP,
 }
 
 CPX_AP_MODULE_ID_LIST = CPX_AP_MODULE_ID_DICT.values()