+++ src/cpx_io/cpx_system/cpx_ap/vaem_ap.py
@@ -1,26 +1,27 @@
-"""CPX-AP-`*`-12DI4DO-`*` module implementation"""
+"""VAEM-L1-S-*-AP module implementation"""
 
 # pylint: disable=duplicate-code
 # intended: modules have similar functions
 
 from cpx_io.cpx_system.cpx_base import CpxBase
+
 from cpx_io.cpx_system.cpx_ap.cpx_ap_module import CpxApModule
-from cpx_io.cpx_system.parameter_mapping import ParameterNameMap
+from cpx_io.cpx_system.cpx_ap import cpx_ap_parameters
 from cpx_io.utils.boollist import bytes_to_boollist, boollist_to_bytes
-from cpx_io.utils.helpers import value_range_check
+from cpx_io.utils.helpers import value_range_check, div_ceil
 from cpx_io.utils.logging import Logging
 from cpx_io.cpx_system.cpx_ap.cpx_ap_enums import (
     LoadSupply,
     FailState,
-    DebounceTime,
 )
 
 
-class CpxAp12Di4Do(CpxApModule):
-    """Class for CPX-AP-`*`-12DI4DO-`*` module"""
+class VaemAP(CpxApModule):
+    """Class for VABX-A-P-EL-E12-AP* module"""
 
     module_codes = {
-        12290: "CPX-AP-A-12DI4DO-M12-5P",
+        8203: "VAEM-L1-S-12-AP",
+        8204: "VAEM-L1-S-24-AP",
     }
 
     def __getitem__(self, key):
@@ -32,47 +33,40 @@ class CpxAp12Di4Do(CpxApModule):
     @CpxBase.require_base
     def read_channels(self) -> list[bool]:
         """read all channels as a list of bool values.
-        Returns a list of 16 elements where the first 12 elements are the input channels 0..11
-        and the last 4 elements are the output channels 0..3
+        Returns a list of all coils
 
         :return: Values of all channels
         :rtype: list[bool]
         """
-        inputs = bytes_to_boollist(self.base.read_reg_data(self.input_register))[:12]
-        outputs = bytes_to_boollist(self.base.read_reg_data(self.output_register))[:4]
-        values = inputs + outputs
+        length = div_ceil(self.information.output_size, 2)
+        data = self.base.read_reg_data(self.output_register, length=length)
+        values = bytes_to_boollist(data, self.information.output_size)
         Logging.logger.info(f"{self.name}: Reading channels: {values}")
         return values
 
     @CpxBase.require_base
-    def read_channel(self, channel: int, output_numbering=False) -> bool:
+    def read_channel(self, channel: int) -> bool:
         """read back the value of one channel
-        Optional parameter 'output_numbering' defines
-        if the outputs are numbered with the inputs ("False", default),
-        so the range of output channels is 12..15 (as 0..11 are the input channels).
-        If "True", the outputs are numbered from 0..3, the inputs cannot be accessed this way.
 
         :param channel: Channel number, starting with 0
         :type channel: int
-        :param output_numbering: Set 'True' if outputs should be numbered from 0 ... 3, optional
-        :type output_numbering: bool
         :return: Value of the channel
         :rtype: bool
         """
-        if output_numbering:
-            channel += 12
-
         return self.read_channels()[channel]
 
     @CpxBase.require_base
     def write_channels(self, data: list[bool]) -> None:
         """write all channels with a list of bool values
 
-        :param data: list of bool values containing exactly 4 elements for each output channel
+        :param data: list of bool values containing exactly the amount of output channels (24/48)
         :type data: list[bool]
         """
-        if len(data) != 4:
-            raise ValueError("Data must be list of four elements")
+        if len(data) != self.information.output_channels:
+            raise ValueError(
+                f"Data must be list of {self.information.output_channels} elements"
+            )
+
         reg = boollist_to_bytes(data)
         self.base.write_reg_data(reg, self.output_register)
 
@@ -87,7 +81,12 @@ class CpxAp12Di4Do(CpxApModule):
         :value: Value that should be written to the channel
         :type value: bool
         """
-        # read current value, invert the channel value
+        if channel not in range(self.information.output_channels):
+            raise ValueError(
+                f"Channel must be in range 0...{self.information.output_channels - 1}"
+            )
+
+        # read current values
         data = bytes_to_boollist(self.base.read_reg_data(self.output_register))
         data[channel] = value
         reg = boollist_to_bytes(data)
@@ -121,36 +120,8 @@ class CpxAp12Di4Do(CpxApModule):
         :type channel: int
         """
         # get the relevant value from the register and write the inverse
-        value = self.read_channel(channel, output_numbering=True)
+        value = self.read_channel(channel)
         self.write_channel(channel, not value)
-
-    @CpxBase.require_base
-    def configure_debounce_time(self, value: DebounceTime | int) -> None:
-        """
-        The "Input debounce time" parameter defines when an edge change of the sensor signal
-        shall be assumed as a logical input signal. In this way, unwanted signal edge changes
-        can be suppressed during switching operations (bouncing of the input signal).
-
-          * 0: 0.1 ms
-          * 1: 3 ms (default)
-          * 2: 10 ms
-          * 3: 20 ms
-
-        :param value: Debounce time for all channels. Use DebounceTime from cpx_ap_enums or
-        see datasheet.
-        :type value: DebounceTime | int
-        """
-
-        if isinstance(value, DebounceTime):
-            value = value.value
-
-        value_range_check(value, 4)
-
-        self.base.write_parameter(
-            self.position, ParameterNameMap()["InputDebounceTime"], value
-        )
-
-        Logging.logger.info(f"{self.name}: Setting debounce time to {value}")
 
     @CpxBase.require_base
     def configure_monitoring_load_supply(self, value: LoadSupply | int) -> None:
@@ -171,7 +142,7 @@ class CpxAp12Di4Do(CpxApModule):
         value_range_check(value, 3)
 
         self.base.write_parameter(
-            self.position, ParameterNameMap()["LoadSupplyDiagSetup"], value
+            self.position, cpx_ap_parameters.LOAD_SUPPLY_DIAG_SETUP, value
         )
 
         Logging.logger.info(f"{self.name}: Setting Load supply monitoring to {value}")
@@ -194,7 +165,7 @@ class CpxAp12Di4Do(CpxApModule):
         value_range_check(value, 2)
 
         self.base.write_parameter(
-            self.position, ParameterNameMap()["FailStateBehaviour"], value
+            self.position, cpx_ap_parameters.FAIL_STATE_BEHAVIOUR, value
         )
 
         Logging.logger.info(f"{self.name}: Setting fail state behaviour to {value}")