+++ tests/unit_tests/cpx_system/cpx_ap/test_cpx_ap4aiui.py
@@ -122,7 +122,7 @@ class TestCpxAp4AiUI:
         cpxap4aiui.base = Mock(write_parameter=Mock())
 
         # Act & Assert
-        with pytest.raises(ValueError):
+        with pytest.raises(IndexError):
             cpxap4aiui.configure_channel_temp_unit(input_value, "None")
 
     @pytest.mark.parametrize(
@@ -195,7 +195,7 @@ class TestCpxAp4AiUI:
         cpxap4aiui.base = Mock(write_parameter=Mock())
 
         # Act & Assert
-        with pytest.raises(ValueError):
+        with pytest.raises(IndexError):
             cpxap4aiui.configure_channel_range(input_value, "None")
 
     @pytest.mark.parametrize(
@@ -370,7 +370,7 @@ class TestCpxAp4AiUI:
         cpxap4aiui.base = Mock(write_parameter=Mock())
 
         # Act & Assert
-        with pytest.raises(ValueError):
+        with pytest.raises(IndexError):
             cpxap4aiui.configure_channel_limits(input_value, upper=0, lower=0)
 
     @pytest.mark.parametrize(
@@ -429,7 +429,7 @@ class TestCpxAp4AiUI:
         cpxap4aiui.base = Mock(write_parameter=Mock())
 
         # Act & Assert
-        with pytest.raises(ValueError):
+        with pytest.raises(IndexError):
             cpxap4aiui.configure_hysteresis_limit_monitoring(input_value, 0)
 
     @pytest.mark.parametrize(
@@ -485,7 +485,7 @@ class TestCpxAp4AiUI:
         cpxap4aiui.base = Mock(write_parameter=Mock())
 
         # Act & Assert
-        with pytest.raises(ValueError):
+        with pytest.raises(IndexError):
             cpxap4aiui.configure_channel_smoothing(input_value, 0)
 
     @pytest.mark.parametrize(
@@ -543,5 +543,5 @@ class TestCpxAp4AiUI:
         cpxap4aiui.base = Mock(write_parameter=Mock())
 
         # Act & Assert
-        with pytest.raises(ValueError):
+        with pytest.raises(IndexError):
             cpxap4aiui.configure_linear_scaling(input_value, True)