+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added 
+- channel_range_check in helpers, now raises IndexError instead of ValueError
+- VMPAL Valve Terminal for CPX-AP
+- VAEM Valve Terminal for CPX-AP
 - Enums for configure functions
 - Systemtests for VABX
 - Examples for cyclic access with threading
@@ -31,6 +34,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Deleted obsolete cpx-e "read_module_count" function
 
 ### Fixed
+- 4DI4DO and 12DI4DO modules toggle channel now working as expected
+- AP Parameter length register is now interpreted correctly as "in bytes"
+- Removed unnecessary validation check in write_parameter_raw
 - Many logger.info strings in modules
 - Docstring in e4aoui configure_channel_range
 - ip-address octetts now in correct order